+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -3,6 +3,7 @@ import ArrowDown from "~/assets/images/icons/arrow-down.svg?react";
 import Bell from "~/assets/images/icons/bell.svg?react";
 import CaretDown from "~/assets/images/icons/caret-down.svg?react";
 import CaretUp from "~/assets/images/icons/caret-up.svg?react";
+import Check from "~/assets/images/icons/check.svg?react";
 import Close from "~/assets/images/icons/close.svg?react";
 import Dashboard from "~/assets/images/icons/dashboard.svg?react";
 import Edit from "~/assets/images/icons/edit.svg?react";
@@ -29,6 +30,7 @@ const iconNameToSvg: Record<
 	bell: Bell,
 	caretDown: CaretDown,
 	caretUp: CaretUp,
+	check: Check,
 	close: Close,
 	dashboard: Dashboard,
 	edit: Edit,