+++ apps/frontend/src/pages/profile/libs/components/user-details/user-details.tsx
@@ -62,20 +62,22 @@ const UserDetails = (): null | React.JSX.Element => {
 	return (
 		<div className={styles["user-details-container"]}>
 			<div className={styles["user-details-avatar"]}>
-				<Avatar size={100} user={authenticatedUser} />
-				<Button
-					label="Upload avatar"
-					onClick={handleTriggerFileUpload}
-					variant="outlined"
-				/>
-				<Button label="Delete avatar" variant="outlined-danger" />
-				<input
-					accept="image/*"
-					onChange={handleFileUpload}
-					ref={fileInputReference}
-					style={{ display: "none" }}
-					type="file"
-				/>
+				<Avatar size={128} user={authenticatedUser} />
+				<div className={styles["user-upload-buttons"]}>
+					<Button
+						label="Upload avatar"
+						onClick={handleTriggerFileUpload}
+						variant="outlined"
+					/>
+					<Button label="Delete avatar" variant="outlined-danger" />
+					<input
+						accept="image/*"
+						className="visually-hidden"
+						onChange={handleFileUpload}
+						ref={fileInputReference}
+						type="file"
+					/>
+				</div>
 			</div>
 			<ProfileForm onSubmit={handleFormSubmit} user={authenticatedUser} />
 		</div>